Adds a migration for a change to the metadata schema EDM-side
This column is not used in the app, however, it should be used as an identifier to assist with explicit sort ordering.

EDM added it while trying to debug some issues. We found that sorting was not the source of the issue.

Instead of reverting the schema change, this commit keeps the migration. It’s necessary still because any schema changes EDM-side will cause the foreign data wrappers to fail, preventing future refreshes.