Re-add silent failure
Previously I made certain calculations explicitly fail, crashing the API. I’m reverting this change because when this error failure actually happens it gets turned into a “null” value in the aggregation. This is what we want.

This was built such that when a formula fails, it defaults to null. This is supposed to capture div/0 handling.