Expand logic for handling change/difference sum/percent calculations in decennial profiles
Previously, decennial change/difference calculations weren’t being made because the code did not take into account that difference in sum calculations may still be performed on decennial profiles withOUT the need for reliability statistics. this is because decennial census data is not sampled and doesn’t include error.

Now, the logic considers the type of profile (decennial) and allows for calculating the change/difference in sum for decennial profiles.

This commit also refactors some of the code is it’s a little easier to follow and exposes the complexity more clearly...